tests + structure update
