+++ qutipy/misc.py
@@ -20,6 +20,35 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 
-from qutipy.misc.base_number_to_int import base_number_to_int
-from qutipy.misc.cvxpy_to_numpy import cvxpy_to_numpy
-from qutipy.misc.numpy_to_cvxpy import numpy_to_cvxpy
+import numpy as np
+from cvxpy import bmat
+
+def base_number_to_int(string, base):
+    b = base
+    string = string[::-1]
+    return sum([string[k] * b**k for k in range(len(string))])
+
+def cvxpy_to_numpy(cvx_obj):
+    """
+    Converts a cvxpy variable into a numpy array.
+    """
+
+    if cvx_obj.is_scalar():
+        return np.array(cvx_obj)
+    elif len(cvx_obj.shape) == 1:  # cvx_obj is a (column or row) vector
+        return np.array(list(cvx_obj))
+    else:  # cvx_obj is a matrix
+        X = []
+        for i in range(cvx_obj.shape[0]):
+            x = [cvx_obj[i, j] for j in range(cvx_obj.shape[1])]
+            X.append(x)
+        X = np.array(X)
+        return X
+
+def numpy_to_cvxpy(np_obj):
+    """
+    Converts numpy array to cvxpy expression.
+    """
+    np_obj_list = np_obj.tolist()
+    return bmat(np_obj_list)
+
