--- qutipy/misc/base_number_to_int.py
@@ -1,32 +1,0 @@
-#               This file is part of the QuTIpy package.
-#                https://github.com/sumeetkhatri/QuTIpy
-#
-#                   Copyright (c) 2022 Sumeet Khatri.
-#                       --.- ..- - .. .--. -.--
-#
-#
-# SPDX-License-Identifier: AGPL-3.0
-#
-#  This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, version 3.
-#
-# This program is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-# General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program. If not, see <http://www.gnu.org/licenses/>.
-#
-
-
-def base_number_to_int(string, base):
-
-    # Last modified: 10 August 2018
-
-    b = base
-
-    string = string[::-1]
-
-    return sum([string[k] * b**k for k in range(len(string))])
